+++ src/Commands.cpp
@@ -263,6 +263,7 @@ static SeqStrings gCommandNames =
     "CmdTocExpandToLevel2\0"
     "CmdTocExpandToLevel3\0"
     "CmdTocCollapseSameLevel\0"
+    "CmdToggleFavoritesSort\0"
     "CmdNone\0"
     "\0";
 
@@ -519,6 +520,7 @@ static i32 gCommandIds[] = {
     CmdTocExpandToLevel2,
     CmdTocExpandToLevel3,
     CmdTocCollapseSameLevel,
+    CmdToggleFavoritesSort,
     CmdNone,
 };
 
@@ -775,6 +777,7 @@ SeqStrings gCommandDescriptions =
     "Bookmarks: Expand to Level 2\0"
     "Bookmarks: Expand to Level 3\0"
     "Bookmarks: Collapse Same Level\0"
+    "Sort Favorites By Name\0"
     "Do nothing\0"
     "\0";
 // clang-format on